feat(logs): nits
Refs: #1667
