+++ packages/api/src/external/commonwell/document/document-query.ts
@@ -63,7 +63,7 @@ import {
   getFileName,
   getContentTypeOrUnknown,
 } from "./shared";
-import { getDocumentReferenceContentTypeCounts } from "../../hie/get-docr-content-type-counts";
+import { getDocumentReferenceContentTypeCounts } from "../../hie/get-counts-analytics";
 import { processAsyncError } from "@metriport/core/util/error/shared";
 
 const DOC_DOWNLOAD_CHUNK_SIZE = 10;
