--- packages/api/src/external/hie/get-docr-content-type-counts.ts
@@ -1,11 +1,0 @@
-export function getDocumentReferenceContentTypeCounts(
-  docRefsContentTypes: string[]
-): Record<string, number> {
-  const contentTypeCounts = docRefsContentTypes.reduce((acc, curr) => {
-    acc[curr] = (acc[curr] || 0) + 1;
-
-    return acc;
-  }, {} as Record<string, number>);
-
-  return contentTypeCounts;
-}
